+++ Cargo.toml
@@ -41,3 +41,4 @@ reqwest = { version = "0.11.23", features = ["json"] }
 sha3 = "0.10.8"
 lazy_static = "1.4.0"
 dotenv = "0.15.0"
+wasm-bindgen = { version = "0.2.91", features = ["serde-serialize"]}
