+++ crates/grapevine_circuits/Cargo.toml
@@ -27,6 +27,8 @@ ff_ce.workspace = true
 hex.workspace = true
 flate2 = "1.0.28"
 
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen.workspace = true
 
 [dev-dependencies]
 criterion = "0.5.1"
